test: avarization
