+++ package.json
@@ -2,16 +2,15 @@
   "name": "kronos-service-consul",
   "version": "0.0.0-semantic-release",
   "description": "kronos service to integrate with consul",
-  "main": "dist/module.js",
-  "module": "src/ServiceConsul.js",
-  "keywords": [
-    "kronos-service"
-  ],
+  "main": "dist/service-consul.js",
+  "module": "src/service-consul.js",
+  "keywords": ["kronos-service"],
   "scripts": {
     "cover": "nyc --temp-directory build/nyc ava",
     "test": "ava",
     "semantic-release": "semantic-release",
-    "docs": "jsdoc2md --configure doc/jsdoc.json -l off -t doc/README.hbs -f src/*.js >README.md",
+    "docs":
+      "jsdoc2md --configure doc/jsdoc.json -l off -t doc/README.hbs -f src/*.js >README.md",
     "pretest": "rollup -c tests/rollup.config.js",
     "posttest": "npm run prepare && markdown-doctest",
     "precover": "rollup -c tests/rollup.config.js",
@@ -62,14 +61,11 @@
   "bugs": {
     "url": "https://github.com/Kronos-Integration/kronos-service-consul/issues"
   },
-  "homepage": "https://github.com/Kronos-Integration/kronos-service-consul#readme",
+  "homepage":
+    "https://github.com/Kronos-Integration/kronos-service-consul#readme",
   "nyc": {
-    "include": [
-      "dist/**/*.js"
-    ],
-    "reporter": [
-      "lcov"
-    ],
+    "include": ["dist/**/*.js"],
+    "reporter": ["lcov"],
     "report-dir": "./coverage"
   },
   "template": {
@@ -81,14 +77,8 @@
     "space": true
   },
   "ava": {
-    "files": [
-      "build/*-test.js"
-    ],
-    "presets": [
-      "env"
-    ],
-    "require": [
-      "babel-register"
-    ]
+    "files": ["build/*-test.js"],
+    "presets": ["env"],
+    "require": ["babel-register"]
   }
 }
